+++ iFruitAddon.csproj
@@ -59,20 +59,20 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="iFruitAddon2.cs" />
-    <Compile Include="Image Types\ContactIcon.cs" />
+    <Compile Include="Textures\ContactIcon.cs" />
     <Compile Include="CustomiFruit.cs" />
-    <Compile Include="iFruitContact.cs" />
-    <Compile Include="iFruitContactCollection.cs" />
-    <Compile Include="Image Types\PhoneImage.cs" />
-    <Compile Include="Image Types\Wallpaper.cs" />
-    <Compile Include="Logger.cs" />
+    <Compile Include="Contacts\iFruitContact.cs" />
+    <Compile Include="Contacts\iFruitContactCollection.cs" />
+    <Compile Include="Textures\PhoneImage.cs" />
+    <Compile Include="Textures\Wallpaper.cs" />
+    <Compile Include="Common\Logger.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Properties\Resources.Designer.cs">
       <AutoGen>True</AutoGen>
       <DesignTime>True</DesignTime>
       <DependentUpon>Resources.resx</DependentUpon>
     </Compile>
-    <Compile Include="Tools.cs" />
+    <Compile Include="Common\Tools.cs" />
   </ItemGroup>
   <ItemGroup>
     <EmbeddedResource Include="Properties\Resources.resx">
